core/rawdb: move usm to meta
