+++ core/rawdb/accessors_chain.go
@@ -20,7 +20,6 @@ import (
 	"bytes"
 	"encoding/binary"
 	"math/big"
-	"time"
 
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/types"
@@ -730,29 +729,3 @@ func FindCommonAncestor(db ethdb.Reader, a, b *types.Header) *types.Header {
 	return a
 }
 
-// ReadUncleanShutdowMarker reads the unclean shutdown marker
-func ReadUncleanShutdowMarker(db ethdb.Reader) (int64, error) {
-	data, err := db.Get(uncleanShutdownPrefix)
-	if err != nil {
-		return 0, err
-	}
-	return int64(binary.BigEndian.Uint64(data)), nil
-}
-
-// WriteUncleanShutdowMarker writes the unclean shutdown marker
-func WriteUncleanShutdowMarker(db ethdb.KeyValueWriter) error {
-	var data = make([]byte, 8)
-	binary.BigEndian.PutUint64(data, uint64(time.Now().Unix()))
-	if err := db.Put(uncleanShutdownPrefix, data); err != nil {
-		log.Warn("Failed to write unclean-shutdown marker", "err", err)
-		return err
-	}
-	return nil
-}
-
-// ClearUncleanShutdowMarker removes the unclean shutdown marker
-func ClearUncleanShutdowMarker(db ethdb.KeyValueWriter) {
-	if err := db.Delete(uncleanShutdownPrefix); err != nil {
-		log.Warn("Failed to remove unclean-shutdown marker", "err", err)
-	}
-}
