+++ les/client.go
@@ -180,7 +180,8 @@ func New(stack *node.Node, config *eth.Config) (*LightEthereum, error) {
 
 	// Check for unclean shutdown
 	if uncleanShutdown, _ := rawdb.ReadUncleanShutdowMarker(chainDb); uncleanShutdown != 0 {
-		log.Error("Unclean shutdown detected", "time", time.Unix(uncleanShutdown, 0))
+		log.Warn("Unclean shutdown detected", "latest", fmt.Sprintf("%v ago",
+			common.PrettyAge(time.Unix(uncleanShutdown, 0))))
 	}
 	// Place new shutdown marker
 	rawdb.WriteUncleanShutdowMarker(chainDb)
